requirements.txt
Django
django-cors-headers
djangorestframework
Pillow
scikit-learn
timm
torchvision
gunicorn